Almost Ready...

+++ pages/api/products/[id].js
@@ -2,7 +2,9 @@ import Product from "../../../models/Product";
 import dbConnect from "../../../util/mongo";
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req;
+    const { method, query: { id }, cookies } = req;
+
+    const token = cookies.token;
 
     dbConnect();
 
@@ -16,6 +18,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'PUT') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json('Unauthorized');
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, { new: true });
             res.status(201).json(product);
@@ -25,6 +30,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'DELETE') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json('Unauthorized');
+        }
         try {
             const product = await Product.findByIdAndDelete(id);
             res.status(200).json("The product is Deleted!");
